Move set iterators to their own file
